multiple constellations per map, multiple maps per invocation
This patch breaks the feature of passing constellations on the command
line, and expects them to be in a list in the options file.  That will
be fixed in a subsequent patch.

A single invocation can now produce multiple maps, as is needed to write
a complete skyculture, and a single map may show multiple
constellations, as is traditional for Ophiuchus+Serpens and
Carina+Puppis+Vela.