chart: ill-fated cache system and other stuff
This patch includes miscellaneous tweaks and a cache system associated
with the chart datatype that will be removed in the next patch.  The
cache system comes from thinking about how to fit a plot to an object,
and draw that object, but not do redundant computation of the object's
projection.  A drawing instructions minilanguage is introduced (or
begun), which will stay.

The program doesn't currently draw stars, nor close loops when drawing
constellations, but those features will be easy enough to bring back
after we get the core drawing protocol settled.