output-skyculture: support charts with arbitrary list of stars to include
..instead of only constellations.

This patch allowed me to make a "big dipper" chart, and was pretty much
the minimum changes to the program needed to do that.  It highlighted the
need for further work on the catalog system.  Catalogs currently return
essentially raw data when queried, so it was necessary to make a constellation
border mimick the properties of stars queried from hyg, so that the fitting
algorithm can process both.  This highlights the need for a system of
well-known datatypes that come out of catalogs - i.e., both stars and
constellation borders are examples of things that have celestial coordinates
so they should have a common interface to that information.

Nevertheless, need to just checkin this work and come back to this issue
later.